Model batch upload does not work for upload graphs
