+++ client/src/services/ModelService.js
@@ -306,6 +306,16 @@ export class ModelService {
     }
   }
 
+  chunkModelsList(array, size) {
+    const chunkedArr = [];
+    let index = 0;
+    while (index < array.length) {
+      chunkedArr.push(array.slice(index, size + index));
+      index += size;
+    }
+    return chunkedArr;
+  }
+
   _addReferencedModels(vertice, sortedModels, checkedList) {
     if (checkedList.some(id => id === vertice.id)) {
       return;
